split register and login views into different files, moved them into a subfolder
