+++ ecommerceapi/views/__init__.py
@@ -1,5 +1,5 @@
 from .payment import PaymentSerializer, Payments
-from .register import register_user, login_user
+from .customer import register_user, login_user
 from .product import Products
 from .order import Orders
 from .order_product import OrderProducts
