+++ ecommerceapi/views/customer/__init__.py
@@ -1,0 +1,2 @@
+from .register import register_user
+from .login import login_user
